Commit 2: refactored changes from merge
