+++ src/main/java/goit/com/shorturlproject/v1/url/UrlLink.java
@@ -1,51 +1,39 @@
-package goit.com.shorturlproject.url;
+package goit.com.shorturlproject.v1.url;
 
-import goit.com.shorturlproject.user.User;
+import goit.com.shorturlproject.v1.user.User;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
 
+@Getter
+@Setter
 @Entity
 @RequiredArgsConstructor
 public class UrlLink {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Getter
-    @Setter
     private long id;
 
     @Column(name = "long_URL")
-    @Getter
-    @Setter
     private String longUrl;
 
     @Column(name = "short_URL")
-    @Getter
-    @Setter
     private String shortUrl;
 
     @Column(name = "created_at")
-    @Getter
-    @Setter
     private LocalDateTime createdAt;
 
     @Column(name = "click_times")
-    @Getter
-    @Setter
     private int clickTimes;
 
     @Column(name = "expiration_date")
-    @Getter
-    @Setter
     private LocalDateTime expirationDate;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
-    @Getter
-    @Setter
     private User user;
 
     @Override
