+++ src/main/java/goit/com/shorturlproject/v1/url/UrlValidator.java
@@ -1,4 +1,4 @@
-package goit.com.shorturlproject.url;
+package goit.com.shorturlproject.v1.url;
 
 import java.net.MalformedURLException;
 import java.net.URL;
