+++ src/main/java/goit/com/shorturlproject/v1/user/User.java
@@ -1,6 +1,6 @@
-package goit.com.shorturlproject.user;
+package goit.com.shorturlproject.v1.user;
 
-import goit.com.shorturlproject.url.UrlLink;
+import goit.com.shorturlproject.v1.url.UrlLink;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +9,8 @@ import lombok.Setter;
 import java.util.HashSet;
 import java.util.Set;
 
+@Getter
+@Setter
 @Entity
 @Table(name = "\"user\"")
 @RequiredArgsConstructor
@@ -16,38 +18,23 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Getter
-    @Setter
     private long id;
 
-    @Getter
-    @Setter
     @Column(name = "first_name")
     private String firstName;
 
-    @Getter
-    @Setter
     @Column(name = "last_name")
     private String lastName;
 
-    @Getter
-    @Setter
     @Column(name = "email")
     private String email;
 
-    @Getter
-    @Setter
     @Column(name = "login")
     private String login;
 
-    @Getter
-    @Setter
     @Column(name = "password")
     private String password;
 
     @OneToMany(mappedBy = "user")
-    @Getter
-    @Setter
     private Set<UrlLink> links = new HashSet<>();
-
 }
