--- src/main/resources/db/migration/local/V1__create-user-for-local-h2.sql
@@ -1,8 +1,0 @@
-CREATE TABLE "user"
-(
-    user_id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    user_name VARCHAR(100) NOT NULL,
-    email VARCHAR(100) NOT NULL,
-    user_login VARCHAR(100) NOT NULL,
-    user_password VARCHAR(100) NOT NULL
-)
