--- src/main/resources/db/migration/prod/V1__create-user-for-prod-postgres.sql
@@ -1,8 +1,0 @@
-CREATE TABLE "user"
-(
-    user_id INTEGER PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
-    username VARCHAR(100) NOT NULL,
-    email VARCHAR(100) NOT NULL,
-    user_login VARCHAR(100) NOT NULL,
-    user_password VARCHAR(100) NOT NULL
-)
